requirements.txt
pipdeptree>=2.0.0,<3.0.0
ansicolors>=1.1.8,<2.0.0
